requirements.txt
CherryPy
path.py
